first refactoring of 3rd checkpoint challenge
