+++ dbc_people.rb
@@ -1,0 +1,5 @@
+module People
+    def offer_high_five
+    	"High five!"
+ 	 end
+end
