+++ senior_teacher.rb
@@ -1,11 +1,15 @@
 require_relative 'people'
+require_relative 'dbc_people'
 
-class SeniorTeacher < People
+class SeniorTeacher < Teachers
   attr_reader :performance_rating, :target_raise
+  include People
+
 
   def initialize(options={})
     super
     @target_raise = 1000
+    @max_rating = 90
   end
 
   def teach_stuff
@@ -13,22 +17,6 @@ class SeniorTeacher < People
     response += "Listen, class, this is how everything works, fo SHO! "
     response += "*drops flat-out insane knowledge bomb* "
     response += "... You're welcome. *saunters away*"
-    response
-  end
-
-  # def receive_raise(raise)
-  #   @salary += raise
-  # end
-
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 90
-      response = "Yay, I'm a great employee!"
-      receive_raise(@target_raise)
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
-    end
     response
   end
